+++ css/premeeting/_connection-status.scss
@@ -4,6 +4,7 @@
     font-size: 12px;
     letter-spacing: 0.16px;
     line-height: 16px;
+    position: absolute;
     width: 100%;
 
     &-header {
